chore: Refactor bags presentation delegate in integration example.
